# specs/openapi.yaml
openapi: 3.1.0
info:
  title: GateLoop API
  version: 1.0.0
  description: >
    REST API for the GateLoop agentic coding platform.
    Every endpoint reflects the live trace; terminal states are first-class.
paths:
...
  /skills/{id}:
    get:
      operationId: getSkill
      summary: STORY-032.4 — a skill's full contents (metadata + SKILL.md + scripts source); static asset, read-only, touches no trace
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Skill metadata, markdown, and scripts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SkillView' }
        '404':
          description: Skill not found
...
components:
  schemas:
...
    SkillView:
      type: object
      description: STORY-032.4 — a skill's full contents (static asset, read-only)
      required: [id, metadata, skill_md, scripts, static_config_level]
      properties:
        id: { type: string }
        metadata: { type: object }
        skill_md: { type: string }
        scripts:
          type: array
          items:
            type: object
            required: [name, source]
            properties:
              name: { type: string }
              source: { type: string }
        static_config_level: { type: boolean, enum: [true] }
...